Remove table editor class to its own file.
